pyproject.toml
[tool.poetry]
name = "wordle"
version = "0.1.0"
description = ""
authors = ["lifehackett"]
readme = "README.md"
# packages = [{include = "wordle", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
pytest = "^7.2.0"
pyyaml = "^6.0"
pytest-env = "^0.8.1"

[tool.poetry.scripts]
wordle = "src.wordle.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
env = [
    "DATASTORE_PATH=./tests/wordle/datastore.yaml"
]
